Add files via upload
(a) Create a html web page that contains a button with which it is
possible to change the text that is shown on the screen. The file
ButtonDemo.html has a JavaScript function named change_text()
which is called after the button is pressed. When the button is
pressed repeatedly the text changes Hello! ... Well done! ... Hello!
... Well done! ... Hello! ..
(b) Modify the program so that the initial text shown on the screen
is &quot;Monday&quot;, and it will change in the following way when the
button is pressed repeatedly: Monday ... Tuesday ... Wednesday
... Thursday ... Friday ... Saturday ... Sunday ... Monday ...
Tuesday ... etc. You should also change the button text so that
there is written &quot;Change day&quot; on the button.