Add files via upload
Create static table with 5 rows and 5 columns with some intial data
and Write a JavaScript function that accept row, column, (to
identify a particular cell) and a string to update the content of that
cell. If the given row and column is not in the range of 5 x 5, it has
to show error message.
Note: For all above html pages style should be given in an external
style sheet and has to be linked to the html page, same style sheet
may be used for all the programs.